movie_mingle_vibe_project/MovieMingle/requirements.txt
Flask
requests
redis
Flask-WTF
python-dotenv